Add support for player position packets
A test client for sending/receiving player packets has been added
(mpClient), and the server also stores a copy of player positions so it
can carry out its own operations on players (ex. collision detection).
Also, the server now supplies the client with the id given to that
client in response to a request for a connection.